requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b2
click==8.0.1
Django==3.2.4
django-cors-headers==3.7.0
django-use-email-as-username==1.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
whitenoise==5.2.0
python-slugify==5.0.2
